Working on screenshot based testing for Korge. pt. 4
